timescaledb-sink-ypqz/requirements.txt
quixstreams[postgresql]==3.14.1
python-dotenv
quixstreams
psycopg2-binary
